backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mongoengine==0.27.0
django-cors-headers==4.3.1
PyJWT==2.8.0
python-decouple==3.8
pandas==2.1.3
django-ratelimit==4.1.0